move container extensions to shared projects
