+++ src/Containers/Prism.DryIoc.Shared/DryIocContainerExtension.cs
@@ -1,0 +1,177 @@
+﻿using System;
+using System.Linq;
+using DryIoc;
+using Prism.Ioc;
+
+namespace Prism.DryIoc
+{
+    /// <summary>
+    /// The <see cref="IContainerExtension" /> Implementation to use with DryIoc
+    /// </summary>
+    public class DryIocContainerExtension : IContainerExtension<IContainer>
+    {
+        /// <summary>
+        /// The instance of the wrapped container
+        /// </summary>
+        public IContainer Instance { get; }
+
+        /// <summary>
+        /// Constructs a new <see cref="DryIocContainerExtension" />
+        /// </summary>
+        /// <param name="container">The <see cref="IContainer" /> instance to use.</param>
+        public DryIocContainerExtension(IContainer container)
+        {
+            Instance = container;
+            Instance.RegisterInstanceMany(new[]
+            {
+                typeof(IContainerExtension),
+                typeof(IContainerProvider)
+            }, this);
+        }
+
+        /// <summary>
+        /// Used to perform any final steps for configuring the extension that may be required by the container.
+        /// </summary>
+        public void FinalizeExtension() { }
+
+        /// <summary>
+        /// Registers an instance of a given <see cref="Type"/>
+        /// </summary>
+        /// <param name="type">The service <see cref="Type"/> that is being registered</param>
+        /// <param name="instance">The instance of the service or <see cref="Type" /></param>
+        /// <returns>The <see cref="IContainerRegistry" /> instance</returns>
+        public IContainerRegistry RegisterInstance(Type type, object instance)
+        {
+            Instance.RegisterInstance(type, instance);
+            return this;
+        }
+
+        /// <summary>
+        /// Registers an instance of a given <see cref="Type"/> with the specified name or key
+        /// </summary>
+        /// <param name="type">The service <see cref="Type"/> that is being registered</param>
+        /// <param name="instance">The instance of the service or <see cref="Type" /></param>
+        /// <param name="name">The name or key to register the service</param>
+        /// <returns>The <see cref="IContainerRegistry" /> instance</returns>
+        public IContainerRegistry RegisterInstance(Type type, object instance, string name)
+        {
+            Instance.RegisterInstance(type, instance, serviceKey: name);
+            return this;
+        }
+
+        /// <summary>
+        /// Registers a Singleton with the given service and mapping to the specified implementation <see cref="Type" />.
+        /// </summary>
+        /// <param name="from">The service <see cref="Type" /></param>
+        /// <param name="to">The implementation <see cref="Type" /></param>
+        /// <returns>The <see cref="IContainerRegistry" /> instance</returns>
+        public IContainerRegistry RegisterSingleton(Type from, Type to)
+        {
+            Instance.Register(from, to, Reuse.Singleton);
+            return this;
+        }
+
+        /// <summary>
+        /// Registers a Singleton with the given service and mapping to the specified implementation <see cref="Type" />.
+        /// </summary>
+        /// <param name="from">The service <see cref="Type" /></param>
+        /// <param name="to">The implementation <see cref="Type" /></param>
+        /// <param name="name">The name or key to register the service</param>
+        /// <returns>The <see cref="IContainerRegistry" /> instance</returns>
+        public IContainerRegistry RegisterSingleton(Type from, Type to, string name)
+        {
+            Instance.Register(from, to, Reuse.Singleton, serviceKey: name);
+            return this;
+        }
+
+        /// <summary>
+        /// Registers a Transient with the given service and mapping to the specified implementation <see cref="Type" />.
+        /// </summary>
+        /// <param name="from">The service <see cref="Type" /></param>
+        /// <param name="to">The implementation <see cref="Type" /></param>
+        /// <returns>The <see cref="IContainerRegistry" /> instance</returns>
+        public IContainerRegistry Register(Type from, Type to)
+        {
+            Instance.Register(from, to);
+            return this;
+        }
+
+        /// <summary>
+        /// Registers a Transient with the given service and mapping to the specified implementation <see cref="Type" />.
+        /// </summary>
+        /// <param name="from">The service <see cref="Type" /></param>
+        /// <param name="to">The implementation <see cref="Type" /></param>
+        /// <param name="name">The name or key to register the service</param>
+        /// <returns>The <see cref="IContainerRegistry" /> instance</returns>
+        public IContainerRegistry Register(Type from, Type to, string name)
+        {
+            Instance.Register(from, to, serviceKey: name);
+            return this;
+        }
+
+        /// <summary>
+        /// Resolves a given <see cref="Type"/>
+        /// </summary>
+        /// <param name="type">The service <see cref="Type"/></param>
+        /// <returns>The resolved Service <see cref="Type"/></returns>
+        public object Resolve(Type type)
+        {
+            return Instance.Resolve(type);
+        }
+
+        /// <summary>
+        /// Resolves a given <see cref="Type"/>
+        /// </summary>
+        /// <param name="type">The service <see cref="Type"/></param>
+        /// <param name="name">The service name/key used when registering the <see cref="Type"/></param>
+        /// <returns>The resolved Service <see cref="Type"/></returns>
+        public object Resolve(Type type, string name)
+        {
+            return Instance.Resolve(type, serviceKey: name);
+        }
+
+        /// <summary>
+        /// Resolves a given <see cref="Type"/>
+        /// </summary>
+        /// <param name="type">The service <see cref="Type"/></param>
+        /// <param name="parameters">Typed parameters to use when resolving the Service</param>
+        /// <returns>The resolved Service <see cref="Type"/></returns>
+        public object Resolve(Type type, params (Type Type, object Instance)[] parameters)
+        {
+            return Instance.Resolve(type, args: parameters.Select(p => p.Instance).ToArray());
+        }
+
+        /// <summary>
+        /// Resolves a given <see cref="Type"/>
+        /// </summary>
+        /// <param name="type">The service <see cref="Type"/></param>
+        /// <param name="name">The service name/key used when registering the <see cref="Type"/></param>
+        /// <param name="parameters">Typed parameters to use when resolving the Service</param>
+        /// <returns>The resolved Service <see cref="Type"/></returns>
+        public object Resolve(Type type, string name, params (Type Type, object Instance)[] parameters)
+        {
+            return Instance.Resolve(type, name, args: parameters.Select(p => p.Instance).ToArray());
+        }
+
+        /// <summary>
+        /// Determines if a given service is registered
+        /// </summary>
+        /// <param name="type">The service <see cref="Type" /></param>
+        /// <returns><c>true</c> if the service is registered.</returns>
+        public bool IsRegistered(Type type)
+        {
+            return Instance.IsRegistered(type);
+        }
+
+        /// <summary>
+        /// Determines if a given service is registered with the specified name
+        /// </summary>
+        /// <param name="type">The service <see cref="Type" /></param>
+        /// <param name="name">The service name or key used</param>
+        /// <returns><c>true</c> if the service is registered.</returns>
+        public bool IsRegistered(Type type, string name)
+        {
+            return Instance.IsRegistered(type, name);
+        }
+    }
+}
