+++ src/Containers/Prism.DryIoc.Shared/Prism.DryIoc.Shared.projitems
@@ -1,0 +1,15 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
+    <HasSharedItems>true</HasSharedItems>
+    <SharedGUID>6e7ec81d-da39-4c4f-a898-0148558c34f4</SharedGUID>
+  </PropertyGroup>
+  <PropertyGroup Label="Configuration">
+    <Import_RootNamespace>Prism.DryIoc</Import_RootNamespace>
+  </PropertyGroup>
+  <ItemGroup>
+    <Compile Include="$(MSBuildThisFileDirectory)DryIocContainerExtension.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)PrismIocExtensions.cs" />
+  </ItemGroup>
+</Project>
