--- src/Forms/Prism.Unity.Forms/PrismIocExtensions.cs
@@ -1,18 +1,0 @@
-﻿using Prism.Ioc;
-using Unity;
-
-namespace Prism.Unity
-{
-    public static class PrismIocExtensions
-    {
-        public static IUnityContainer GetContainer(this IContainerProvider containerProvider)
-        {
-            return ((IContainerExtension<IUnityContainer>)containerProvider).Instance;
-        }
-
-        public static IUnityContainer GetContainer(this IContainerRegistry containerRegistry)
-        {
-            return ((IContainerExtension<IUnityContainer>)containerRegistry).Instance;
-        }
-    }
-}
