+++ src/Prism.Core/Ioc/IContainerRegistry.cs
@@ -2,22 +2,75 @@
 
 namespace Prism.Ioc
 {
+    /// <summary>
+    /// The registering container
+    /// </summary>
     public interface IContainerRegistry
     {
+        /// <summary>
+        /// Registers an instance of a given <see cref="Type"/>
+        /// </summary>
+        /// <param name="type">The service <see cref="Type"/> that is being registered</param>
+        /// <param name="instance">The instance of the service or <see cref="Type" /></param>
+        /// <returns>The <see cref="IContainerRegistry" /> instance</returns>
         IContainerRegistry RegisterInstance(Type type, object instance);
 
+        /// <summary>
+        /// Registers an instance of a given <see cref="Type"/> with the specified name or key
+        /// </summary>
+        /// <param name="type">The service <see cref="Type"/> that is being registered</param>
+        /// <param name="instance">The instance of the service or <see cref="Type" /></param>
+        /// <param name="name">The name or key to register the service</param>
+        /// <returns>The <see cref="IContainerRegistry" /> instance</returns>
         IContainerRegistry RegisterInstance(Type type, object instance, string name);
 
+        /// <summary>
+        /// Registers a Singleton with the given service and mapping to the specified implementation <see cref="Type" />.
+        /// </summary>
+        /// <param name="from">The service <see cref="Type" /></param>
+        /// <param name="to">The implementation <see cref="Type" /></param>
+        /// <returns>The <see cref="IContainerRegistry" /> instance</returns>
         IContainerRegistry RegisterSingleton(Type from, Type to);
 
+        /// <summary>
+        /// Registers a Singleton with the given service and mapping to the specified implementation <see cref="Type" />.
+        /// </summary>
+        /// <param name="from">The service <see cref="Type" /></param>
+        /// <param name="to">The implementation <see cref="Type" /></param>
+        /// <param name="name">The name or key to register the service</param>
+        /// <returns>The <see cref="IContainerRegistry" /> instance</returns>
         IContainerRegistry RegisterSingleton(Type from, Type to, string name);
 
+        /// <summary>
+        /// Registers a Transient with the given service and mapping to the specified implementation <see cref="Type" />.
+        /// </summary>
+        /// <param name="from">The service <see cref="Type" /></param>
+        /// <param name="to">The implementation <see cref="Type" /></param>
+        /// <returns>The <see cref="IContainerRegistry" /> instance</returns>
         IContainerRegistry Register(Type from, Type to);
 
+        /// <summary>
+        /// Registers a Transient with the given service and mapping to the specified implementation <see cref="Type" />.
+        /// </summary>
+        /// <param name="from">The service <see cref="Type" /></param>
+        /// <param name="to">The implementation <see cref="Type" /></param>
+        /// <param name="name">The name or key to register the service</param>
+        /// <returns>The <see cref="IContainerRegistry" /> instance</returns>
         IContainerRegistry Register(Type from, Type to, string name);
 
+        /// <summary>
+        /// Determines if a given service is registered
+        /// </summary>
+        /// <param name="type">The service <see cref="Type" /></param>
+        /// <returns><c>true</c> if the service is registered.</returns>
         bool IsRegistered(Type type);
 
+        /// <summary>
+        /// Determines if a given service is registered with the specified name
+        /// </summary>
+        /// <param name="type">The service <see cref="Type" /></param>
+        /// <param name="name">The service name or key used</param>
+        /// <returns><c>true</c> if the service is registered.</returns>
         bool IsRegistered(Type type, string name);
     }
 }
