--- src/Wpf/Prism.DryIoc.Wpf/Ioc/PrismIocExtensions.cs
@@ -1,18 +1,0 @@
-﻿using DryIoc;
-using Prism.Ioc;
-
-namespace Prism.DryIoc
-{
-    public static class PrismIocExtensions
-    {
-        public static IContainer GetContainer(this IContainerProvider containerProvider)
-        {
-            return ((IContainerExtension<IContainer>)containerProvider).Instance;
-        }
-
-        public static IContainer GetContainer(this IContainerRegistry containerRegistry)
-        {
-            return ((IContainerExtension<IContainer>)containerRegistry).Instance;
-        }
-    }
-}
