+++ src/Wpf/Prism.DryIoc.Wpf/Prism.DryIoc.Wpf.csproj
@@ -13,11 +13,13 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="DryIoc.dll" Version="4.0.7" />
+    <PackageReference Include="DryIoc.dll" Version="4.1.4" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\Prism.Wpf\Prism.Wpf.csproj" />
   </ItemGroup>
+
+  <Import Project="..\..\Containers\Prism.DryIoc.Shared\Prism.DryIoc.Shared.projitems" Label="Shared" />
 
 </Project>
