+++ src/Wpf/Prism.Unity.Wpf/PrismApplication.cs
@@ -1,7 +1,5 @@
 ﻿using System;
 using Prism.Ioc;
-using Prism.Regions;
-using Prism.Unity.Ioc;
 using Unity;
 
 namespace Prism.Unity
